pyproject.toml
[tool.poetry]
name = "GenPlanner"
version = "0.1.0"
license = "BSD-3-Clause"
description = "kek"
authors = ["DDonnyy <63115678+DDonnyy@users.noreply.github.com>"]
readme = "README.md"

packages = [{ include = "src", from = "gen_planner/python" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
tqdm = "^4.66.2"
numpy = "^2.1.3"
pandas = "^2.2.0"
geopandas = "^1.0.0"
loguru = "^0.7.2"
scipy = "^1.13.1"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"
gunicorn = "^23.0.0"
geojson-pydantic = "^1.1.2"



[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pylint = "^3.0.3"
isort = "^5.13.2"
jupyter = "^1.0.0"
maturin = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.pylint.format]
max-line-length = 120
expected-line-ending-format = "LF"
max-locals = 20
extension-pkg-allow-list = ["networkit"]
disable = [
    "duplicate-code",
    "missing-module-docstring",
    "missing-function-docstring",
    "too-many-locals",
    "too-many-branches",
    "too-many-statements",
    "too-many-arguments",
    "cyclic-import",
    "line-too-long"
]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
split_on_trailing_comma = true
skip = ["__init__.py"]
